move helper functions to `utils.rs`
