+++ src/lib.rs
@@ -201,6 +201,7 @@ mod file_dialog;
 /// Information panel showing the preview and metadata of the selected item
 pub mod information_panel;
 mod modals;
+mod utils;
 
 pub use config::{
     FileDialogConfig, FileDialogKeyBindings, FileDialogLabels, FileDialogStorage, IconFilter,
